Semester project files
EXAMINATION PROCESSING SYSTEM
This project requires you to work in groups of three (5) students to create a C sharp program that can process examination for a university department. More specifically, the program should:
1.	Read the details of at least ten (10) students. Such details include student name and ID Number.
2.	Read the overall score (marks) in each subject. The program should allow score entry in at least 5 subjects.
3.	Calculate the average grade based on the USIU grading system.
4.	Display a report card on the screen showing:
a.	The name of the student
b.	The ID Number of the student
c.	All the subjects and their scores
d.	The average score and the grade attained.
e.	A recommendation (e,g good, poor, excellent etc)
f.	The date the report is being viewed.

Note:
•	You are encouraged to use functions for each of the tasks 1 to 4
•	This project will be marked in week 13 of the semester.